Added validation

+++ DAL/Implementation/Repositories/CrewRepository.cs
@@ -1,8 +1,10 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using DAL.Interfaces;
 using DAL.Models;
 using Microsoft.EntityFrameworkCore;
+using Shared.Exceptions;
 
 namespace DAL.Implementation.Repositories
 {
@@ -29,13 +31,28 @@ namespace DAL.Implementation.Repositories
 
         public void Create(Crew entity)
         {
+            if (entity == null)
+            {
+                throw new ArgumentNullException(nameof(entity));
+            }
+            
             context.Crews.Add(entity);
         }
 
         public void Update(Crew entity)
         {
+            if (entity == null)
+            {
+                throw new ArgumentNullException(nameof(entity));
+            }
+            
             var oldEntity = context.Crews.Include(c => c.Pilot).Include(c => c.Stewardesses)
                 .FirstOrDefault(c => c.Id == entity.Id);
+            if (oldEntity == null)
+            {
+                throw new NotFoundException(nameof(oldEntity));
+            }
+            
             context.Entry(oldEntity).State = EntityState.Detached;
             context.Entry(entity).State = EntityState.Modified;
         }
@@ -43,7 +60,11 @@ namespace DAL.Implementation.Repositories
         public void Delete(int id)
         {
             var entity = context.Crews.Find(id);
-
+            if (entity == null)
+            {
+                throw new NotFoundException(nameof(entity));
+            }
+            
             context.Crews.Remove(entity);
         }
     }
