+++ DAL/Implementation/Repositories/DepartureRepository.cs
@@ -1,9 +1,11 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using DAL.Interfaces;
 using DAL.Models;
 using Microsoft.EntityFrameworkCore;
+using Shared.Exceptions;
 
 namespace DAL.Implementation.Repositories
 {
@@ -36,12 +38,27 @@ namespace DAL.Implementation.Repositories
 
         public void Create(Departure entity)
         {
+            if (entity == null)
+            {
+                throw new ArgumentNullException(nameof(entity));
+            }
+            
             context.Departures.Add(entity);
         }
 
         public void Update(Departure entity)
         {
+            if (entity == null)
+            {
+                throw new ArgumentNullException(nameof(entity));
+            }
+            
             var oldEntity = context.Departures.Find(entity.Id);
+            if (oldEntity == null)
+            {
+                throw new NotFoundException(nameof(oldEntity));
+            }
+            
             context.Entry(oldEntity).State = EntityState.Detached;
             context.Entry(entity).State = EntityState.Modified;
         }
@@ -49,7 +66,11 @@ namespace DAL.Implementation.Repositories
         public void Delete(int id)
         {
             var entity = context.Departures.Find(id);
-
+            if (entity == null)
+            {
+                throw new NotFoundException(nameof(entity));
+            }
+            
             context.Departures.Remove(entity);
         }
     }
