+++ DAL/Implementation/Repositories/StewardessRepository.cs
@@ -1,9 +1,11 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using DAL.Interfaces;
 using DAL.Models;
 using Microsoft.EntityFrameworkCore;
+using Shared.Exceptions;
 
 namespace DAL.Implementation.Repositories
 {
@@ -30,12 +32,27 @@ namespace DAL.Implementation.Repositories
 
         public void Create(Stewardess entity)
         {
+            if (entity == null)
+            {
+                throw new ArgumentNullException(nameof(entity));
+            }
+            
             context.Stewardesses.Add(entity);
         }
 
         public void Update(Stewardess entity)
         {
+            if (entity == null)
+            {
+                throw new ArgumentNullException(nameof(entity));
+            }
+            
             var oldEntity = context.Stewardesses.Find(entity.Id);
+            if (oldEntity == null)
+            {
+                throw new NotFoundException(nameof(oldEntity));
+            }
+            
             context.Entry(oldEntity).State = EntityState.Detached;
             context.Entry(entity).State = EntityState.Modified;
         }
@@ -43,7 +60,11 @@ namespace DAL.Implementation.Repositories
         public void Delete(int id)
         {
             var entity = context.Stewardesses.Find(id);
-
+            if (entity == null)
+            {
+                throw new NotFoundException(nameof(entity));
+            }
+            
             context.Stewardesses.Remove(entity);
         }
     }
